lab3 - add settimeout command

+++ lab3/include/timer.h
@@ -11,8 +11,8 @@ struct timer_event {
 void infinite(void);
 void core_timer_enable(void);
 void core_timer_disable(void);
-void add_core_timer(void (*func)(void*), void *args, unsigned int time);
-void set_core_timer(unsigned int time);
+void add_core_timer(void (*func)(void*), void *args, unsigned long time);
+void set_core_timer(unsigned long time);
 void remove_core_timer(void);
 void print_elapsed_time(void);
 void print_core_timer_message(void *args);
